+++ NEWS.md
@@ -16,10 +16,17 @@
 * feat: TabResNet learner now supports lazy tensors.
 * feat: The `LearnerTorch` base class now supports the private method `$.ingress_tokens(task, param_vals)`
   for generating the `torch::dataset`.
+* fix: `NA` is now a valid shape for lazy tensors
+* feat: lazy tensors of length 0 can now be materialized.
 * feat: `nn("block")` (which allows to repeat the same network segment multiple
   times) now has an extra argument `trafo`, which allows to modify the
   parameter values per layer.
 * feat: Shapes can now have multiple `NA`s and not only the batch dimension can be missing. However, most `nn()` operators still expect only one missing values and will throw an error if multiple dimensions are unknown.
+* feat: Progress callback got argument `digits` for controlling the precision
+  with which validation/training scores are logged.
+* feat: Training now does not fail anymore when encountering a missing value
+  during validation but uses `NA` instead.
+* feat: The context for callbacks now includes the network prediction (`y_hat`).
 
 # mlr3torch 0.2.1